working grid and gulp. next step yeoman
