+++ examples/excelGrid.html
@@ -4,7 +4,7 @@ react grid
 
 <div id="edit-test"></div>
 <div id="example"></div>
-	<script type="text/javascript" src="/build/js/example1.js"></script>
+  <script type="text/javascript" src="/build/js/examples/excelGrid.jsx"></script>
