+++ lib/Canvas.js
@@ -18,9 +18,11 @@
  */
 "use strict";
 
+
+/* jshint esnext: true */
+
 var React          = require('react/addons');
 var shallowEqual   = require('./utils').shallowEqual;
-var cx             = React.addons.classSet;
 var cloneWithProps = React.addons.cloneWithProps;
 var ScrollShim     = require('./ScrollShim');
 var Row            = require('./Row');
