+++ lib/Cell.js
@@ -4,7 +4,6 @@
 'use strict';
 
 var React = require('react/addons');
-var cx    = React.addons.classSet;
 
 var Cell = React.createClass({
   render: function() {
