+++ lib/ColumnMetrics.js
@@ -3,6 +3,7 @@
  */
 "use strict";
 
+/* jshint esnext: true, unused: false, undef: false */
 var React               = require('react');
 var shallowCloneObject  = require('./shallowCloneObject');
 var DOMMetrics          = require('./DOMMetrics');
