+++ lib/DOMMetrics.js
@@ -3,6 +3,9 @@
  */
 'use strict';
 
+
+/* jshint esnext: true, browser: true */
+
 var React = require('react');
 var utils = require('./utils');
