+++ lib/Row.js
@@ -3,6 +3,8 @@
  */
 'use strict';
 
+/* jshint esnext: true */
+
 var React = require('react/addons');
 var cx    = React.addons.classSet;
 var Cell  = require('./Cell');
