+++ lib/Scrollbar.js
@@ -134,6 +134,6 @@ var HorizontalScrollbar = React.createClass({
 });
 
 module.exports = {
-  VerticalScrollbar,
-  HorizontalScrollbar
+  VerticalScrollbar: VerticalScrollbar,
+  HorizontalScrollbar: HorizontalScrollbar
 };
