+++ lib/Viewport.js
@@ -79,31 +79,25 @@ var ViewportScroll = {
   },
 
   updateScroll: function(scrollTop, scrollLeft, height, rowHeight, length) {
-    var renderedRowsCount = ceil(height / rowHeight);
+    var nextScrollState = { };
+      
+    nextScrollState.renderedRowsCount = ceil(height / rowHeight);
 
-    var visibleStart = floor(scrollTop / rowHeight);
+    nextScrollState.visibleStart = floor(scrollTop / rowHeight);
 
-    var visibleEnd = min(
-        visibleStart + renderedRowsCount,
+    nextScrollState.visibleEnd = min(
+        nextScrollState.visibleStart + nextScrollState.renderedRowsCount,
         length);
 
-    var displayStart = max(
+    nextScrollState.displayStart = max(
         0,
-        visibleStart - renderedRowsCount * 2);
+        nextScrollState.visibleStart - nextScrollState.renderedRowsCount * 2);
 
-    var displayEnd = min(
-        visibleStart + renderedRowsCount * 2,
+    nextScrollState.displayEnd = min(
+        nextScrollState.visibleStart + nextScrollState.renderedRowsCount * 2,
         length);
 
-    var nextScrollState = {
-      visibleStart,
-      visibleEnd,
-      displayStart,
-      displayEnd,
-      height,
-      scrollTop,
-      scrollLeft
-    };
+
 
     this.setState(nextScrollState);
   },
