+++ lib/getScrollbarSize.js
@@ -1,5 +1,7 @@
 "use strict";
 
+/* jshint browser: true */
+
 var size;
 
 function getScrollbarSize() {
