+++ lib/getWindowSize.js
@@ -10,6 +10,7 @@
  *
  * @return {Object} height and width of the window
  */
+ /* jshint browser: true */
 function getWindowSize() {
     var width = window.innerWidth;
     var height = window.innerHeight;
@@ -24,7 +25,7 @@ function getWindowSize() {
         height = document.body.clientHeight;
     }
 
-    return {width, height};
+    return {width: width, height: height};
 }
 
 module.exports = getWindowSize;
