+++ lib/utils.js
@@ -2,8 +2,9 @@
  * @jsx React.DOM
  */
 "use strict";
+var utils = {};
 
-function mergeInto(dst, src) {
+utils.mergeInto = function mergeInto(dst, src) {
   if (src != null) {
     for (var k in src) {
       if (!src.hasOwnProperty(k)) {
@@ -12,16 +13,16 @@ function mergeInto(dst, src) {
       dst[k] = src[k];
     }
   }
-}
+};
 
-function merge(a, b) {
+utils.merge = function merge(a, b) {
   var result = {};
-  mergeInto(result, a);
-  mergeInto(result, b);
+  utils.mergeInto(result, a);
+  utils.mergeInto(result, b);
   return result;
-}
+};
 
-function shallowEqual(a, b) {
+utils.shallowEqual = function shallowEqual(a, b) {
   if (a === b) {
     return true;
   }
@@ -42,19 +43,19 @@ function shallowEqual(a, b) {
   }
 
   return true;
-}
+};
 
-function emptyFunction() {
+utils.emptyFunction = function emptyFunction() {
 
-}
+};
 
-function invariant(condition, message) {
+utils.invariant = function invariant(condition, message) {
   if (!condition) {
     throw new Error(message || 'invariant violation');
   }
-}
+};
 
-function shallowCloneObject(obj) {
+utils.shallowCloneObject = function shallowCloneObject(obj) {
   var result = {};
   for (var k in obj) {
     if (obj.hasOwnProperty(k)) {
@@ -62,13 +63,6 @@ function shallowCloneObject(obj) {
     }
   }
   return result;
-}
-
-module.exports = {
-  shallowEqual,
-  emptyFunction,
-  invariant,
-  shallowCloneObject,
-  mergeInto,
-  merge
 };
+
+module.exports = utils;
